requirements.txt
fastapi==0.111.0
motor==3.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
aiosmtplib==3.0.1
google-generativeai==0.5.4
apscheduler==3.10.4
cryptography==42.0.7
python-dotenv==1.0.1
dnspython==2.6.1
bleach==6.1.0
slowapi==0.1.9
pydantic-settings==2.2.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.1
gspread==6.0.2
aiofiles==23.2.1
